requirements.txt
streamlit
pandas
sentence-transformers
faiss-cpu
langchain
google-generativeai
python-dotenv 